clean up of results and actions
