+++ src/InputText.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 class TextInput extends Component {
   constructor(props) {
     super(props)
-    console.log(props)
+
     this.state = {
       value: props.value || ''
     }
