+++ src/search/Results.js
@@ -1,9 +1,5 @@
 import React, { Component } from 'react'
-import { Link } from 'react-router-dom'
-
-import { nestStateForApi } from '../utils/convert'
-
-import Alert from '../Alert'
+import ResultsActions from './ResultsActions'
 
 import './Results.css'
 
@@ -12,59 +8,22 @@ class SearchResults extends Component {
     super(props)
 
     this.state = {
-      httpError: null
+      error: null
     }
 
     this.tables = new Map()
-    this.buttons = new Map()
 
-    this.handleViewMoreClick = this.handleViewMoreClick.bind(this)
-    this.handleDeleteClick = this.handleDeleteClick.bind(this)
-    this.toggleAreYouSure = this.toggleAreYouSure.bind(this)
+    this.removeAnInstitutionFromState = this.removeAnInstitutionFromState.bind(
+      this
+    )
   }
 
-  toggleAreYouSure(key) {
-    this.setState({ httpError: null })
-    document.getElementById(`initialActions${key}`).classList.toggle('hidden')
-    document.getElementById(`areYouSure${key}`).classList.toggle('hidden')
-  }
-
-  handleDeleteClick(institution, key, token) {
-    fetch('/v2/admin/institutions', {
-      method: 'DELETE',
-      body: JSON.stringify(nestStateForApi(institution)),
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-        Authorization: 'Bearer ' + token
-      }
-    })
-      .then(response => {
-        if (response.ok) {
-          return response.json()
-        } else {
-          throw new Error(response.status)
-        }
-      })
-      .then(json => {
-        // need to remove the institution from the state
-        this.props.deleteAnInstitution(key)
-      })
-      .catch(error => {
-        this.setState({ httpError: error.message })
-      })
-  }
-
-  handleViewMoreClick(i) {
-    const table = this.tables.get(i)
-    const button = this.buttons.get(i)
-
-    table.classList.toggle('hidden')
-    if (table.classList.contains('hidden')) {
-      button.innerHTML = 'Show other fields'
-    } else {
-      button.innerHTML = 'Hide other fields'
-    }
+  removeAnInstitutionFromState(key) {
+    let newInstitutions = this.state.institutions.filter(
+      (institution, i) => i !== key
+    )
+    if (newInstitutions.length === 0) newInstitutions = null
+    this.setState({ institutions: newInstitutions })
   }
 
   // TODO: make this a component
@@ -76,32 +35,6 @@ class SearchResults extends Component {
       <h2>
         {numOfResults} {resultsText} found
       </h2>
-    )
-  }
-
-  renderAction(institution) {
-    let link = {
-      pathname: '/update',
-      text: 'Update',
-      type: 'update'
-    }
-
-    if (institution.activityYear === 2017) {
-      link = {
-        pathname: '/add',
-        text: 'Add',
-        type: 'addition'
-      }
-    }
-
-    return (
-      <React.Fragment>
-        <Link
-          to={{ pathname: link.pathname, state: { institution: institution } }}
-        >
-          {link.text}
-        </Link>
-      </React.Fragment>
     )
   }
 
@@ -133,51 +66,13 @@ class SearchResults extends Component {
                     <td>{institution.respondentName}</td>
                     <td>{institution.emailDomains}</td>
                     <td>{institution.taxId}</td>
-
-                    <td className="action">
-                      <div className="initialActions" id={`initialActions${i}`}>
-                        <button
-                          onClick={event => this.handleViewMoreClick(i)}
-                          ref={element => this.buttons.set(i, element)}
-                          className="showOtherFields"
-                        >
-                          Show other fields
-                        </button>
-                        {this.renderAction(institution)}
-                        <button
-                          className="delete"
-                          onClick={event => this.toggleAreYouSure(i)}
-                        >
-                          Delete
-                        </button>
-                      </div>
-                      <div className="areYouSure hidden" id={`areYouSure${i}`}>
-                        <span>Are you sure?</span>{' '}
-                        <button
-                          className="delete"
-                          onClick={event =>
-                            this.handleDeleteClick(
-                              institution,
-                              i,
-                              this.props.token
-                            )}
-                        >
-                          Yes
-                        </button>
-                        <button onClick={event => this.toggleAreYouSure(i)}>
-                          No
-                        </button>
-                      </div>
-                      {this.state.httpError ? (
-                        <Alert
-                          type="error"
-                          heading="Access Denied"
-                          text="Sorry, it doesn't look like you have the correct permissions to
-                perform this action."
-                          smallWidth={true}
-                        />
-                      ) : null}
-                    </td>
+                    <ResultsActions
+                      institution={institution}
+                      i={i}
+                      error={this.state.error}
+                      handleDeleteClick={this.props.handleDeleteClick}
+                      tables={this.tables}
+                    />
                   </tr>
                   <tr
                     className="otherData hidden"
